jinja2 pre processing and some minor changes
You can now add additional html snippets which will be integrated in
jinja2 templates. This processing is done in cb_start: we'll load our layouts
and do a simple substition of {{ include: identifier }}, where identifier will
be replaced by the corresponding value in `env`. See layouts/main.html and
layouts/entry.html for an example implementation of Disqus.

All links in templates are now relative, which makes them independent of
server-side HTTP oder HTTPS. Unfortunately, disqus requires a absolute path,
which is used, when disqus is enabled; same for syndication.