close #8 using a tree views -> filters
With this optimization re-generation of the whole blog without
cache is 4-5 times faster (3.66 secs vs. 15.21 secs on my machine).

It results in a few more files in .cache/ and should change in some
way, later. Previously there have been around 2.7 cache files per
entry, now it slighty grows to around 2.9/3.0 (tested with 147
and five different views).

This commit also introduces a new API for custom filters and is not
finished yet. First, `__init__` and `__call__` moved to `init` and
`transform`, which should reflect their purpose. Furthermore, `init`
is now invoked before the actual Filter object is created. When the
filter is actually created it receives the function name and arguments
hence, init should be used to initialize only acrylamid independend
things.