close #6 again, it's a fix for built-in filters
Since jinja2 compiles to native python code, a filter is required
in the current template context and must not defaulted to lambda x: x.
The fix only applies to all built-in filters and is a hardcoded
filter, that does nothing by default and should be overwritten in the
context method in the view.