new flag --ignore to continue compile on (import-)errors
I have changed the behavior when a built-in filter requires a module
which is not available (and would cause an ImportError normally). Now
if you don't have this module installed and don't use this filter (or
markdown extension) it won't bother you.

If you require this filter/markdown-extension it either raises an
exception an aborts (this avoids having a clean cache but containing
incorrectly rendered HTML) or you can pass -i, --ignore to ignore
errors but might have issues with cached and therefore not re-compiled
items.