add filter version, close #26
When a filter fixes an issue, acrylamid automatically re-compiles
all affected entries.

This fixes also a regression introduced with d6dc8f6dc4 that did
ignore modifications in the array of filters (in given context)
so that an article gets marked as *has_changed*.

Now an article is marked as has_changed when one or more intermediates
(including the final result) are missing, the cache file does not exist
or the mtime of entry.filename > cache's filename.