fix regression introduced with 4b5188ba
Using HTMLParser rather than regular expressions makes currently
no difference in speed but later we can easier add an fast xml
parser than a regular expression engine.

This fixes an issue where we could not resolve a pattern to it's
long description. Unfortunately this is a little bit slower now.