move code of loading entries to readers.load
Now it's just `load(conf)` and you have your entrylist. By default
we recognize .txt, .rst and .md extensions (they seem to be very
common as filename extension) but as "fallback" we test wether
a found file is binary or text and use it as well if it's text.

By this time I renamed `ENTRIES_IGNORE` to `CONTENT_IGNORE` (a list
of fnmatch patterns).