add support for translated posts, closes #22
To enable the Translation view, add this to your :doc:`conf.py`:

.. code-block:: python

    '/:year/:slug/:lang/': {
        'view': 'translation',
        'template': 'main.html',  # default, includes entry.html
    }

Translations are posts with the same `identifier` and a different `lang` attribute.
An example:

The English article::

    ---
    title: Foobar is not dead
    identifier: foobar-is-not-dead
    ---

    That's true, foobar is still alive!

And the French version::

    ---
    title: Foobar n'est pas mort !
    identifier: foobar-is-not-dead
    lang: fr
    ---

    Oui oui, foobar est toujours vivant !

If the blog language is ``"en"`` then the english article will be included into
the default listing but the french version not. You can link to the translated
versions via:

.. code-block:: html+jinja

    {% if 'translation' in env.views and env.translationsfor(entry) %}
    <ul>
        {% for tr in env.translationsfor(entry) %}
            <li><strong>{{ tr.lang }}:</strong>
                <a href="{{ env.path ~ tr.permalink }}">{{ tr.title }}</a>
            </li>
        {% endfor %}
    </ul>
    {% endif %}