get length of datetime.timedelta via total_seconds()
My machine is running directly in UTC timezone.
The calculation '(datetime.now() - datetime.utcnow())' returns a
'datetime.timedelta()' object with '-1 day, 23:59:59.999948' which
should work.  But when getting the '.seconds' attribute of this object
the value '86399' is returned. This are ~1440 minutes which are 24
hours. Those are then passed as an offset. This breaks the function
'tzinfo.utcoffset()' with the following exception:

'ValueError: tzinfo.utcoffset() returned 1440; must be in -1439 .. 1439'

Simply changing '.seconds' to '.total_seconds()' fixed the issue for me.
This returns '5.2e-05' which gets rounded down to 0.