*real* theming support, part of #62
Only two noteworthy changes:

- LAYOUT_DIR was renamed to THEME (no deprecation work-around)
- the default stylesheets are now located inside the theme directory

The current theme structure (you can read the documentation in
`docs/theming.rst`) is as follows:

acrylamid/defaults/
  - sample-entry.txt
  - conf.py
  - mako|jinja2
    - atom.xml
    - rss.xml
  - html5
    - jinja2
    - mako
  - xhtml
    - jinja2

All themes are re-using stuff that already exists (conf.py, sample-entry.txt,
rss and atom xml). To achieve this, a theme requires a short __init__.py that
returns a mapping of destination of a file or bunch of files.