Change the ID for the atom feed itself so it is unique.
Especially when there are several feeds, e.g. when using a feed for
each tag with the AtomPerTag view, you need unique IDs for each feeds.

I chose to use the same format as for the IDs of the entries. I took
the date of the first commit of the Acrylamid repository.