add archive listing filter
.. code-block:: html+jinja2

    {% for year in env.globals.entrylist | archivesfor %}
    <h2>{{ year ~ ' (' ~ year | count ~ ')' }}</h2>
    <ul>
        {% for month in year %}
        <li>
            <a href="{{ env.path ~ '/' ~ year ~ '/' ~ month ~ '/'  }}">{{ month.full }}</a>
            ({{  month | count }})
        </li>
        {% endfor %}
    </ul>
    {% endfor %}

This generates a listing that shows the amount of postings during the
period. You can also iterate over each month to get to the days.
:class:`Year`, :class:`Month` and :class:`Day` objects always evaluate to
a zero-padded date unit such as 2012 (year) or 01 (January. In addition,
Month and Day objects have ``full`` and ``abbr`` attributes to access the
fullname or abbreviation in your current location.