include filter configuration namespace into hash ...
... and allow the inclusion of certain environment/configuration
variables into the hash, too. For example could you include the
modification timestamp of an external file, so Acrylamid automatically
updates the content (for example within the acronym filter).

The environment has now a new attribute `.filter` that contains the
actual used filters. This way, the environment hash depends on the
filter hash as well and can trigger re-compilation under given
circumstances.