Fixed bug https://github.com/posativ/acrylamid/issues/251
For some reason if you had multiple types of entries (say two completely
seperate blogs (blog1 and blog2), the permalinks for *all* blogs would get set
to either /blog1/* or /blog2/* (with the /blog{1,2}/:slug/ view rules).

This fixes it so the permalink for a given entry, if not already set, is
assigned correctly.

The only changes occured in acrylamid/views/entry.py.

It was basically a one-line fix.